New content with pinning & changing <h1>s
Three new sections have been added (BufferZone.js, InevitableWar.js, & Outnumbered.js) with descriptive paragraphs describing the various stages leading up to the start of the Winter War. The <h1>s change from Finnish/Russian to English as the user scrolls down the page. A pre-war map of Finnish Karelia has a pinning effect that allows the user to view the map for a longer period of time as they read through the text. Lastly, the headers in Intro.js are now in the middle of the screen as opposed to the bottom.