Fixed pinning & positioning issues
Previously inside divs with the image to the left, the ScrollMagic pinning would push the right-aligned paragraph too far down the page, & an extremely negative value (i.e. -80vh) had to be given to the paragraph to push it back up. Now, all ScrollMagic pinning scenes have the setting pushFollowers: false. This makes it so the pinned image doesn't affect the positioning of the paragraph. Now all paragraphs have regular positioning from the top of the page. Additionally, the navbar close symbol (x) has been changed to an up arrow, & the id has been changed accordingly in the stylesheet & js file. Lastly, a tooltip was added to the down arrow telling the user to scroll down, & another link has been added to Outnumbered.js that provides more info on the Red Army.