React-particles-js update + new image hover effects
New particle effects have been added to all divs with green backgrounds (.buffer-zone-div, .it-begins-div, .the-beginning-of-the-end-div & .footer-div). These particle effects simulate snow & take place in the background, allowing the user to still interact with all images & a tags within the div. This was implemented using react-particles-js, an react version of Vincent Garreau's particles.js on npm. Additionally, both #menu-white & #up-arrow-green now have the same hover effect that all a tags do. This was done by making clones of these images & triggering the clone's animation (which changes the opacity) on hover. These clones are positioned 1px down & 1px to the right of the original image.