Gradient backgrounds & numerous other small tweaks
The solid green backgrounds in several divs have been replaced by gradient green backgrounds. The code for this background is stored in a new mixin called green-gradient-background. Also, <Particles/> components have been moved out of the app.js file into each file where it is used. The # of particles as well as their speed has been upped slightly, & the position of the .snow-background canvar selector has been changed from fixed to absolute. The background image for .home-field-advantage-div has been changed to increase text legibility & a link in the footer has been removed because it already exists on another part of the page. The <a> tags in the navbar now have smaller horizontal margins so they fit more easily in the navbar, & the text for <a> tags that take the user offsite has been simplified throughout the page. Lastly, Bootstrap values have been adjusted to only take place on larger screens, & additional scss has been added in preparation for making the page responsive.