Image hover events now handled by React
React is now used to manage <img> onHover events. In every case, onMouseEnter is used on the outer div to trigger a state change, which then sets the style of inner <img> to brightness: 30% & the style of the inner <p> to opacity: 1. Likewiese, in every image onMouseLeave is used on the outer div to trigger a stage change which undos the aforementioned effects. All css that was formerly used to take care of this has been removed. Lastly, a new component called Menu.js has been created which contains #menu, #menu-clone, & openNavigation().